Ship a more sane .gitattributes with the project
This is related to the issue contributors had earlier on with the line endings. This solution basically forbids line-ending changes on all files by default and only marks specific files as "text" which then means that the local settings of the person checking the files out will determine what happens to the line endings on her machine.